fix: read upload CSP from runtime environment
